requirements.txt
google-generativeai>=1.21.1
requests>=2.32.4
python-dotenv>=1.0.0
